* reorganize the tests directory structure
