+++ tests/core/data/test_data/test_config_ray.yaml
@@ -4,7 +4,7 @@ project_name: 'ray-demo-new-config'
 dataset:
   configs:
     - type: local
-      path: ./data/sample.jsonl  # path to your dataset directory or file
+      path: ./test_data/sample.jsonl  # path to your dataset directory or file
 
 export_path: './outputs/demo/demo-processed'
